# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Treasury API
  version: 0.2.0
  description: Transfers, reserves, holds, reconciliation, and settlement operations
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/internal/recon/sources/{id}:
    get:
      summary: Get recon source
      description: Returns a single reconciliation source by ID.
      x-platform:
        kind: resource
        entity: ReconSource
        affectsLedger: false
      operationId: getReconSource
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Reconciliation source details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": &a4
          description: Resource not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update recon source
      description: Updates a reconciliation source.
      x-platform:
        kind: resource
        entity: ReconSource
        affectsLedger: false
      operationId: updateReconSource
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                config:
                  type: object
                metadata:
                  type: object
      responses:
        "200":
          description: Reconciliation source updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
    delete:
      summary: Delete recon source
      description: Deletes a reconciliation source by ID.
      x-platform:
        kind: resource
        entity: ReconSource
        affectsLedger: false
      operationId: deleteReconSource
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Reconciliation source deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deleted
                  - id
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
        "401": *a1
        "404": *a4
  /api/internal/recon/sources/{id}/imports:
    get:
      summary: List imports for source
      description: Returns a paginated list of imports for a reconciliation source.
      x-platform:
        kind: resource
        entity: ReconImport
        affectsLedger: false
      operationId: listReconImports
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of imports.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
        "404": *a4
    post:
      summary: Create import
      description: Creates a new import for a reconciliation source.
      x-platform:
        kind: resource
        entity: ReconImport
        affectsLedger: false
      operationId: createReconImport
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_url
              properties:
                file_url:
                  type: string
                  format: uri
                format:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Import created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
  /api/internal/recon/sources/{id}/imports/{importId}:
    get:
      summary: Get import
      description: Returns a single import by ID.
      x-platform:
        kind: resource
        entity: ReconImport
        affectsLedger: false
      operationId: getReconImport
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - &a5
          in: path
          name: importId
          required: true
          description: Import ID.
          schema:
            type: string
      responses:
        "200":
          description: Import details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401": *a1
        "404": *a4
    patch:
      summary: Update import
      description: Updates an import record.
      x-platform:
        kind: resource
        entity: ReconImport
        affectsLedger: false
      operationId: updateReconImport
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - *a5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Import updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "404": *a4
  /api/internal/recon/periods:
    get:
      summary: List reconciliation periods
      description: Returns a paginated list of reconciliation periods.
      x-platform:
        kind: resource
        entity: ReconPeriod
        affectsLedger: false
      operationId: listReconPeriods
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of reconciliation periods.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": *a1
        "403": *a2
    post:
      summary: Create recon period
      description: Creates a new reconciliation period.
      x-platform:
        kind: resource
        entity: ReconPeriod
        affectsLedger: false
      operationId: createReconPeriod
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_date
                - end_date
              properties:
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                metadata:
                  type: object
      responses:
        "201":
          description: Reconciliation period created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": *a3
        "401": *a1
        "403": *a2
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...